refactor() : 공식 스타일 가이드에 따라 파일명 - snakecase, 클래스명 - PascalCase로 변경
