+++ lib/dto/link.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-part 'Link.g.dart'; // 이게 있어야 build_runner가 연결 가능
+part 'link.g.dart';
 
 @JsonSerializable()
 class Link{
@@ -19,7 +19,7 @@ class Link{
   final String linkName;
   final String thumbnailUrl;
   final String faviconUrl;
-  // 👇 이 두 줄 반드시 추가!
+
   factory Link.fromJson(Map<String, dynamic> json) =>
       _$LinkFromJson(json);
