+++ lib/provider/folders_provider.dart
@@ -1,12 +1,12 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:krap/dto/Folder.dart';
-import 'package:krap/provider/client_provider.dart';
+import 'package:krap/dto/folder.dart';
+import 'package:krap/provider/repo_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'folders_provider.g.dart';
 
 @riverpod
 Future<List<Folder>> folders(Ref ref, int id) async {
-  final repo = ref.watch(userRepoProvider);
-  return repo.getUser(id);
+  final repo = ref.watch(folderRepoProvider);
+  return repo.getFolders(id);
 }
